requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.9
webargs==8.0.1
joblib==1.3.2
binance-connector==3.4.0
binance-futures-connector==3.3.1
